+++ test/routes/actions.test.js
@@ -1,10 +1,10 @@
 const ApplicationContext = require('../../src/context/application-context');
 const ActionsRoutes = require('../../src/routes/actions');
 
-function initContext(schema) {
+function initContext(schema, smartActionHookGetResponse) {
   const context = new ApplicationContext();
   context.init((ctx) => ctx
-    .addInstance('logger', { warn: jest.fn() })
+    .addInstance('logger', { warn: jest.fn(), error: jest.fn() })
     .addInstance('pathService', {
       generate: jest.fn((path) => path),
       generateForSmartActionCustomEndpoint: jest.fn((path) => path),
@@ -13,11 +13,46 @@ function initContext(schema) {
       parameterize: jest.fn((name) => name),
     })
     .addInstance('schemasGenerator', { schemas: schema })
+    .addInstance('smartActionHook', { getResponse: smartActionHookGetResponse })
     .addClass(ActionsRoutes)
     .addValue('model', { name: 'users' })
     .addValue('implementation', { getModelName: jest.fn((m) => m.name) })
     .addValue('app', { post: jest.fn() }));
   return context;
+}
+
+async function callHook(hooks, smartActionHookGetResponse, requestBody) {
+  const schemas = {
+    users: {
+      actions: [{
+        name: 'send invoice',
+        hooks,
+        fields: [{ field: 'invoice number', type: 'String' }],
+      }],
+    },
+  };
+  const {
+    actions, model, app, logger,
+  } = initContext(schemas, smartActionHookGetResponse).inject();
+
+  const request = { body: requestBody || { recordsId: [1] } };
+  const send = jest.fn((values) => values);
+  const response = { status: jest.fn(() => ({ send })) };
+  const perform = jest.fn(() => ({ id: 1, name: 'Jane' }));
+  const implementation = {
+    getModelName: jest.fn((m) => m.name),
+    ResourceGetter: jest.fn(() => ({ perform })),
+  };
+
+  await actions.perform(app, model, implementation, {}, {});
+
+  const [, , callback] = app.post.mock.calls[0];
+
+  await callback(request, response);
+
+  return {
+    send, response, model, implementation, logger,
+  };
 }
 
 describe('routes > actions', () => {
@@ -36,7 +71,7 @@ describe('routes > actions', () => {
     expect(app.post).not.toHaveBeenCalled();
   });
 
-  it('should not create a route when no actions.values is present', async () => {
+  it('should not create a route when actions.values or actions.hooks.* are missing', async () => {
     expect.assertions(4);
 
     const schema = { users: { actions: [{}, {}] } };
@@ -123,6 +158,212 @@ describe('routes > actions', () => {
       expect(response.status).toHaveBeenCalledTimes(1);
       expect(send).toHaveBeenCalledTimes(1);
       expect(result).toStrictEqual({ name: 'Jane' });
+    });
+  });
+
+  describe('when action.hooks is present', () => {
+    describe('when action.hooks.load is present', () => {
+      it('should create a route', async () => {
+        expect.assertions(4);
+
+        const schema = { users: { actions: [{ name: 'send invoice', hooks: { load: jest.fn() } }] } };
+        const {
+          actions, pathService, stringUtils, model, implementation, app,
+        } = initContext(schema).inject();
+
+        await actions.perform(app, model, implementation, {}, {});
+
+        expect(stringUtils.parameterize).toHaveBeenCalledTimes(1);
+        expect(pathService.generate).toHaveBeenCalledTimes(1);
+        expect(app.post).toHaveBeenCalledTimes(1);
+
+        const [path] = app.post.mock.calls[0];
+        expect(path).toBe('actions/send invoice/hooks/load');
+      });
+
+      describe('when calling the route controller', () => {
+        it('should call the load hook service', async () => {
+          expect.assertions(2);
+
+          const load = jest.fn();
+          const smartActionHookGetResponse = jest.fn();
+          const { model, implementation } = await callHook({ load }, smartActionHookGetResponse);
+
+          expect(implementation.ResourceGetter).toHaveBeenNthCalledWith(1, model, { recordId: 1 });
+          expect(smartActionHookGetResponse).toHaveBeenNthCalledWith(
+            1,
+            load,
+            [{ field: 'invoice number', type: 'String' }],
+            { id: 1, name: 'Jane' },
+          );
+        });
+
+        it('should fail with message when load hook service throws', async () => {
+          expect.assertions(2);
+
+          const load = jest.fn();
+          const smartActionHookGetResponse = jest.fn(() => { throw new Error('oops'); });
+          const { send, response } = await callHook({ load }, smartActionHookGetResponse);
+
+          expect(response.status).toHaveBeenNthCalledWith(1, 500);
+          expect(send).toHaveBeenNthCalledWith(1, { message: 'oops' });
+        });
+
+        it('should succeed with the updated fields when load hook service response', async () => {
+          expect.assertions(2);
+
+          const newFields = [{ field: 'invoice number', type: 'String', value: 'hello from load' }];
+
+          const load = jest.fn();
+          const smartActionHookGetResponse = jest.fn(() => newFields);
+          const { send, response } = await callHook({ load }, smartActionHookGetResponse);
+
+          expect(response.status).toHaveBeenNthCalledWith(1, 200);
+          expect(send).toHaveBeenNthCalledWith(1, { fields: newFields });
+        });
+      });
+    });
+    describe('when action.hooks.change is present', () => {
+      it('should create a route', async () => {
+        expect.assertions(4);
+
+        const schema = { users: { actions: [{ name: 'send invoice', hooks: { change: { foo: jest.fn() } } }] } };
+        const {
+          actions, pathService, stringUtils, model, implementation, app,
+        } = initContext(schema).inject();
+
+        await actions.perform(app, model, implementation, {}, {});
+
+        expect(stringUtils.parameterize).toHaveBeenCalledTimes(1);
+        expect(pathService.generate).toHaveBeenCalledTimes(1);
+        expect(app.post).toHaveBeenCalledTimes(1);
+
+        const [path] = app.post.mock.calls[0];
+        expect(path).toBe('actions/send invoice/hooks/change');
+      });
+
+      describe('when calling the route controller', () => {
+        it('should send undefined to hook service when change field is unreachable', async () => {
+          expect.assertions(1);
+
+          const smartActionHookGetResponse = jest.fn();
+          await callHook(
+            { change: { 'this field does not exist': jest.fn() } },
+            smartActionHookGetResponse,
+            { recordsId: [1], fields: [{ field: 'invoice number', type: 'String' }] },
+          );
+
+          expect(smartActionHookGetResponse).toHaveBeenNthCalledWith(
+            1,
+            null,
+            [{ field: 'invoice number', type: 'String' }],
+            { id: 1, name: 'Jane' },
+          );
+        });
+
+        it('should send undefined to hook service when no field has previousValue and value different', async () => {
+          expect.assertions(1);
+
+          const smartActionHookGetResponse = jest.fn();
+          const field = {
+            field: 'foo',
+            type: 'String',
+            previousValue: 'a',
+            value: 'a',
+          };
+          await callHook(
+            { change: { foo: jest.fn() } },
+            smartActionHookGetResponse,
+            {
+              recordsId: [1],
+              fields: [field],
+            },
+          );
+
+          expect(smartActionHookGetResponse)
+            .toHaveBeenNthCalledWith(1, null, [field], { id: 1, name: 'Jane' });
+        });
+
+        it('should call the change hook service', async () => {
+          expect.assertions(2);
+
+          const smartActionHookGetResponse = jest.fn();
+          const field = {
+            field: 'foo',
+            type: 'String',
+            previousValue: 'a',
+            value: 'b',
+          };
+          const change = { bar: jest.fn(), foo: jest.fn(), baz: jest.fn() };
+          const { implementation, model } = await callHook(
+            { change },
+            smartActionHookGetResponse,
+            {
+              recordsId: [1],
+              fields: [field],
+            },
+          );
+
+          expect(implementation.ResourceGetter).toHaveBeenNthCalledWith(1, model, { recordId: 1 });
+          expect(smartActionHookGetResponse)
+            .toHaveBeenNthCalledWith(1, change.foo, [field], { id: 1, name: 'Jane' });
+        });
+
+        it('should fail with message when change hook service throws', async () => {
+          expect.assertions(2);
+
+          const smartActionHookGetResponse = jest.fn(() => { throw new Error('oops'); });
+          const field = {
+            field: 'foo',
+            type: 'String',
+            previousValue: 'a',
+            value: 'b',
+          };
+          const change = { bar: jest.fn(), foo: jest.fn(), baz: jest.fn() };
+          const { send, response } = await callHook(
+            { change },
+            smartActionHookGetResponse,
+            {
+              recordsId: [1],
+              fields: [field],
+            },
+          );
+
+          expect(response.status).toHaveBeenNthCalledWith(1, 500);
+          expect(send).toHaveBeenNthCalledWith(1, { message: 'oops' });
+        });
+
+        it('should succeed with the updated fields when change hook service response', async () => {
+          expect.assertions(2);
+
+          const newFields = [{
+            field: 'invoice number',
+            type: 'String',
+            value: 'hello from load',
+            previousValue: 'a',
+          }];
+
+          const smartActionHookGetResponse = jest.fn(() => newFields);
+          const field = {
+            field: 'foo',
+            type: 'String',
+            previousValue: 'a',
+            value: 'b',
+          };
+          const change = { bar: jest.fn(), foo: jest.fn(), baz: jest.fn() };
+          const { send, response } = await callHook(
+            { change },
+            smartActionHookGetResponse,
+            {
+              recordsId: [1],
+              fields: [field],
+            },
+          );
+
+          expect(response.status).toHaveBeenNthCalledWith(1, 200);
+          expect(send).toHaveBeenNthCalledWith(1, { fields: newFields });
+        });
+      });
     });
   });
 });